src/api/users/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers


class UserTokenSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('username', 'first_name', 'last_name', 'email')
