src/requirements.txt
Django==3.2
djangorestframework==3.11.0
django-environ==0.9.0
django-grappelli==3.0.3
django-treenode==0.18.1
django-extensions==3.1.5

psycopg2-binary==2.9.3
drf-yasg==1.20.0

gunicorn==20.1.0
whitenoise==6.2.0
django-cors-headers==3.13.0

# pytest==7.2.1
# pytest-cov==4.0.0
# django-debug-toolbar==3.2
# flake8==3.9.0
# pre-commit==2.11.1

